LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-nested-cross-in-sandboxed-cross-url-block.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="../../resources/frame-ancestors-test.js"></script>
</head>
<body>
<script>
description("A 'frame-ancestors' CSP directive with a URL value should compare against each frame's origin rather than URL, " +
            "so a nested frame with a sandboxed parent frame should be blocked due to the parent having a unique origin.");

testNestedIFrame(SAMEORIGIN_ORIGIN + " " + CROSSORIGIN_ORIGIN, CROSS_ORIGIN, CROSS_ORIGIN, EXPECT_BLOCK, "allow-scripts");
</script>
</body>
</html>
